+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Fix a bug when retried background data migration can not start
 - Do not run multiple background schema migrations on the same table at the same time
 
 ## 0.17.1 (2024-04-28)